Initial experiment with LLVM IR
This is an experiment to see if I can compile LLVM IR for getting and
setting the stack location. From here, I will experiment with using the
clang/gcc builtins __builtin_setjmp() and __builtin_longjmp(). If
successful, this should give us the building blocks for CPU-only
coroutines in Rust with only a tiny bit of LLVM IR and C.