Make it impossible to mishandle a cancelation
Before this change if the parent canceled the child then the child could
still return a value. This wouldn't cause a crash, but was clearly a
logic error. After this patch this is impossible.

We solve this problem by creating a Finished wrapper which cannot be
constructed outside the crate. This wrapper is used in all closure
signatures, enforcing that the returned value must be either Canceled or
Finished. Neither type can be constructed by the closure. Therefore, if
the coroutine is canceled the only logical possiblity is to return the
Canceled type.