check for early drop of Coroutine and handle dirty stack
check for drop of the coroutine before yield()
check for drop of the corouting within done()

For the check to work, we need the Context<Y,R> part of the stack zeroed before
using it, so use MaybeUninit::zeroed() instead of MaybeUninit::uninit().

Also updated the tests to contain non-zero stack pages.