Potential fix for the stackrestore problem
By coincindence llvm.stackrestore works on x86. The other architectures
suffer from the compiler not knowing to use the old stackpointer to
load the parameters for %func() and even the value for %func.

The correct handling would be to load those values in registers before
calling llvm.stackrestore and use the to call %func.

This workaround uses alloca to store and load them, which on some
architectures (tested on armv7l) works.